pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "huggingface-hub>=0.34.4",
    "instructorembedding>=1.0.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.29",
    "openai>=1.99.9",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
    "tiktoken>=0.11.0",
]
